+++ Sources/KeyboardKit/_Deprecated/CalloutStyles.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardCalloutStyle")
+public typealias CalloutStyle = KeyboardCalloutStyle
+
+@available(*, deprecated, renamed: "KeyboardActionCalloutStyle")
+public typealias ActionCalloutStyle = KeyboardActionCalloutStyle
+
+@available(*, deprecated, renamed: "KeyboardInputCalloutStyle")
+public typealias InputCalloutStyle = KeyboardInputCalloutStyle